Change of focus
I was having a block. I was not sure where to start or how to
state clearly what I wanted to achieve. Somehow I came up with
this simple scheme: ask yourself what is it that you want out of
this software? Or what is the problem that the software is trying
to solve. I realized that these are two angles that one can use to
define software: what are the things that the software does but
also what are the things that the software helps you avoid. Kind
of like a tug-of-war.

This method reminds me of one of Rich Hickey's lecture at Clojure
Conj 2023 tittled "Design in Practice". In this lecture he
mentioned something about keeping track of things in an Excel file.
At the time, I found it a little inconvenient and restricting to
do things using a rubric, but after working on Fyznos I realized
that design is a multistep process and the initial steps are
always simple and always the same, regardless of the product you
are making. At least that is what I think. In this case, since
this application is really simple and it is addressing a single
problem, I think that this methodology is a good start. Something
like this would not work for Fyznos, since it does not have a
clear problem to solve. That is the reason why I am using
principle-based specification inspired by constitutional law
(which I admit is a really odd source of inspiration, particularly
for the dystopian fumbles of modern democracies ... but such is
the nature of inspiration).

As I was writting the previous specs, I realized that there were
some edge cases that I was not considering. I realized that I need
to have some examples on what kind of transformations I want to
perform. This problem does fit that simple type of testing. For
Fyznos, on the other hand, it is going to be much trickier to
test. For the next step I am going to create some examples as
tests and do some TDD before I try any implementation.