Deleting case-two
In this snippet I found a problematic use of the em-dash, which in
markdown (as far as I know) is represented with a regular slash
(-). I think this makes a fine edge case.

Here I need to choose whether to make a special rule to detect the
em-dash and its usage or make the user responsible for properly
formating it, in order to keep the algorithm simple. I decided to
do the second rather than the first. That is - in regards to the
em-dash - the user must leave empty spaces around the dash to
indicate the em-dash and consider hyphenated words like em-dash,
matter-of-factly, short-term, full-scale, or old-fashion a single
word.

For words where the user places a dash at the beggining or at the
end by mistake, like -this or this- the program would consider
them a whole word and let the user fix them on their own. Once the
user makes the fix, the program will rearrange words as needed on
the next run and output the desired result.