Fith case + verification
This case was longer than the other and I messed up in the middle
of writting the output. I pressed \ instead of enter. The I hit
Ctrl+z an extra time and it undid the line previous to that one.
This made me realize that it was really simple to make a mistake
to the output. For TDD that is based on value, having wrong values
is a serious concern. I need to make sure that the tests are
correct. With wrong tests, TDD is counter productive.

To combat this, I decided to use the good-old do it twice method
to verify something. Now all I need to do is write a utility that
verifies that the test is a valid one. I also thought of verifying
that the input is correct. I could as easily make an error while
copying/modifying an input file. Since the variations between input
and output are in the location of whitespaces and newlines, then
we can remove them and check that the input and output files
contain the same symbols. That way we can make sure that no
character of interest was deleted nor that a stowaway character
was introduced by accident.

The question is, should each test be verified every time is run or
should this varification be external? The main point of unit
testing is to be fast. Ideally is should take at most a second,
two is ok, and more than 5 is streching it. Thus, I would like the
tool to run independently. But then that would require devs to run
it every time they modify/add test cases. Human error could creep
in. In the event that a test fails, devs would be left wondering
it there is an error in the test. They might have to check if the
test is valid, so why not test it anyway? Thus, we should check if
the test is valid before we run it.

Or what if we automate the way I described earlier. What if we
check the validity of the test after it fails? Most of the times
devs get their code right. Thus mosts test runs would not have any
overhead. And the devs would not have any doubts about the
validity of the tests nor would they need to remember to run
something after they are done modifying/adding test cases. The
only issue with this scheme is that every time a test fails there
is going to be overhead.

The verification of a test is only needed once. The verification
only needs to be run again if the file is modified. Hashes are a
common way to keep track if a file has been modified. There could
be a way to keep track of whether a verification has been run for
the current files used in the test. That should be faster to run.
Thus, at the beggining of every test, the test suit would check to
see if any test case needs to be verified and performs the
varifications as needed. Or, using the idea that verifications are
only needed when test fails, the former verification process could
be run only when a test fails. That way most of the time there is
no overhead. But then it could be the case that if there is a long
period of times without a failed test and too many files are
modified, the moment that there is a failing test, there would be
a large overhead. Unless we only verify the failing tests and
leave the other pending. We could combine both strategies by using
randomization. Randomly the test could decide to run the
verification process to reduce the chances of large queues and we
keep a local verification process on test fails. That would mean
that from time to time a test would have overhead. What if instead
of having a full verification process, we only verify one pending
test case? In that case, from time to time there would be minimal
overhead.

I think it is a double edge sword for tests to have internal state.
The argument could also be made that this is the reason why TDD is
not a good idea. This program is a simple markdown formater. I am
sure there are better formaters out there. Just look at how much I
am (potentially) going to implement to verify my program. What are
the odds that I would mess up my implementation? Wouldn't it be
faster to just verify it manually? Or fix bugs as they are found
by users? For the time being I am going to architect the solutions
that I described previously just as a form of practice but this
shows the economic drawbacks of TDD. And here is the kicker ...
these verifications are not proofs. Despite all of this effort the
users might still encounter errors. So, is it really worth it?

I almost forgot but I also came up with some property-based
tests.  The current scheme consists of running

  symbolsOnly(input) == symbolsOnly(output)      and
              output == verification

but the property that we wish to see is that the distance between
the last character in a line and the 100th character position is
smaller than the length of the first word in the next line
(assuming there is a next line). Therefore, what we seek is:

                                      d
                                   |-----|
                                   V     V
  Dog ... the last word in this line\n   * <- 100th char position
  cat ... sternocleidomastoid in this line\n

In this example we want the distance d from char 'e' and the 100th
char position to be smaller than the first word of the next line
(if there is one). We could make it so that the distance function
is possitive if the last char is before the 100th char, zero if it
is at 100th char, and negative if it is after the 100th char. Then
we could define a passing tests as a file where every line gives a
positive distance and where the distance of each line is smaller
than the length of the first word in the next line, with the
exception of the last line.

Making software feels like an accident. Had I not taken the time
to think about the (input, output) tests, would I have been able
to figure out the property-based testing that I just described?
Perhaps I would have, but would that had happened before I
started to develop the software? Software specification takes
time. But no matter how hard you think about things, in many
ways, the result is an accident. There is an uncanny percentage
of good and bad things that happened due to chance. How do you
make a living off of something so finicky?

Also, note to self. Add a newline char at the end of the file to
make it easier to modify and to account for that pesky but gedit
has. It doesn't want to show the last line of the file, for some
weird reason.