Adding verifications and drafting algo
I worked on the verification for two test cases. To create the
verification files, I copied the input file and modified it to
its correct state. As I was working on the second verificationb, I
realized that it was a good opportunity to try to craft the
algorithm to perform the program. At first I was trying to deduce
the algorithm logically, but I opted for something more annecdotal.
I decided to write down how I was modifying the file and tried to
figure out the reasoning behind my actions.

Thanks to this exercise I was able to identify the components of
interest in the problem. That is, what are the conditions that
determine how the program should behave. Up until that point, the
exercise was useful and relatively quick (writing things down
slowed me down - but that was not strictly necessary). After I had
identified all of the moving parts, I decided to try and piece
together the algorithm that I was already making in my head. I
tried to resist comming up with the algorithm on my head, on my
own, without having all of the facts. I believe there is such a
thing as a methodology to crafting algoriths. That is, a
metaheuristic. Yes, algorithm design is an art, but artists also
have stablished methods - like paining the background before the
foreground. Similarly, for this problem I thought that a state
machine was the appropriate approach. Why? It was just a gut
feeling. I just thought that it resembled the problems that I
did when working with Turing Machines.

But as I was working on the algorithm, I quickly ran into the
limitations of trying to describe a state machine using English.
It is possible, but it gets tricky. In particular, it was taking
too much time. I am concerned about my speed as a developer. These
days companies prefer to be faster in the market wrong and fix
things as they come as opposed to making software right from the
beggining. Thus, only methods of verification that are cost
effective are used. No one is going to bother reading something
like the dev-journal that I created. That thing is only useful as
a temp file, not as something that needs to be preserved. It is
something that bothers me but it is the reality of how this
industry works.

But in particular, what bother me about my algorithm is that it is
all about how to react to the current state of the algorithm, but
has nothing to do with the descriptions made during algorithm
discovery. That is, I am describing when to do swaps between
newline and whitespace chars but I am not describing that a line
trim or a line merge was happening. That is what is bothering me.
I think I am overstepping the algorithm. The goal of this exercise
is to come up with a methodology, not just to solve a single
problem.

It was taking too much time to document the algoritm discovery
process. But just trying to do a trial an error with just code is
also something that I consider risky. But that is not the case
with this project, since there are tests. This made me think of
what role tests play in this scenario. Without tests, a trial an
error is likely to ignore edge cases. As I was working on the
verifications, I realized that there were alot of cases covered:
when the 101th char is a whitepsace, then the line has a lot of
space for mergin the next line, where there is an empty line, etc.
Since I have such extensive tests, doing a trial and error
approach isn't so risky. Add to that the property-based test and
it should be ok to just attempt to solve the problem and let the
tests guide development.

All I need to do is to be very articulate during the algorithm
discovery using a trial and error method.

The issue with that, though, is that good tests do not equal good
diagnostics. The tests will indicate that something went wrong but
not what/why it went wrong. How to speed such thing?

Conclusion:
With good test cases and good naming/abstraction conventions there
is no need to document algorithm discovery/development. The use
of manual execution helps in identifying important components of
the algorithm and it is encouraged, but there is no need to
document it. Fast diagnostics for failing tests is still an open
problem.