Embarrasing typo
Yep. It was am embarrasing typo. Its nil, not nill.
Not only that. The test passed ... which it shouldn't do.

PASS
ok	github.com/gonzalez-nunez-uriel/w100	0.001s

And what happened to the output I was expecting? Did the go
compiler ignore the function because the output was not used or
something?

Actually, it did not log anything. There should have been a
start/end message for the tests. What happened?