Manual failing
I did not know how to trigger the log. The only thing I could do
was change the test from passing to failing. It was bothering me
that the test was passing without implementation. I miss Scala's
not implemented feature. Anyway, I decided to fail it manually for
now ... and the logs appeared. I guess it makes sense, since the
whole point is to run tests as fast as possible.

Thus, logs only print in the even of a test fail. Print debugging
is definetely not the way to do this, but I have no idea how else
to do it. I miss a REPL on which to test things.