First internal function
I decided to implement the symbolsOnly function in an internal
package. The reasoning is simple. i did not wanted my test file to
grow too much but I did not wanted otehr to be able to import that
package. As I understand, that kind of code goes in the 'internal'
directory.

I think I implemented it just fine ... but go test cannot reach it.
go test only tests the upper level tests. I have to go to the
directory of the package that I am working on to test it. Is that
the way it is supposed to work? I mean, it makes sense ... but I
find it a little weird. I guess is so that only relevant tests are
made. Why would you run the tests of other packages? The child
packages are unaware (and therefore independent) of parent
packages. There is no reason why a modification in a parent
package could affect a child package. The other way around is not
true. I guess I will just have to be careful.

What if I want to run all the tests(just as a diagnostics)?
Go is weird. I am starting to not like it. But it is still the
language that I hate the least.