An error in the test
This shows the pitfalls of value-based testing. It is a silly
mistake. Why even bother with doing this? I didn't saved this in
version control, so I undid it to show how easy it is to make a
mistake.

\# github.com/gonzalez-nunez-uriel/w100/internal/symbols [github.com/gonzalez-nunez-uriel/w100/internal/symbols.test]
./symbols.go:6:2: withoutWhiteSpaces declared and not used
FAIL	github.com/gonzalez-nunez-uriel/w100/internal/symbols [build failed]

In line 7 I used input as the argument to the function instead of
withoutWhiteSpaces. I caught it because the compiler warned me
that a variable was left unused. I am using gedit and I did not
use copy+paste. It was a cognitize dissonance error. I typed the
wrong thing, because ... I don't know. I just fucked up. It
happens to me more than I would like to admit. But it is important
to document these fuckups. I am glad I finally got the jist of how
to use Git. Although I almost missed documenting this.

Property based testing is the best way to go. But the issue still
remains - you could get the property generator wrong. And even
though it is possible to wrap an API behind common properties,
there will always be things that the API cannot handle. Also the
API is also possible to be wrong. Verifying software is tricky and
I am runnig out of time.