Property based test
Using strings.Contains to check if there are any offending chars.
I realized that it would be best to split the tests to get more
granularity in the diagnostics. That is something that I want to
do next. I was planning on having values as well as properties
together just in case one failed, but in this particular case I
don't think it would be very useful. For that reason I am going to
delete them in the next commit.