To close or not to close
I am using os.ReadFile/1. Do I need to close it? The function
implies that all the contents of the file are made available and
from the name I infer that I only want the contents, not the file.
It wouldn't be unreasonable to expect the file handled to be
closed right after all of the contents have been extracted.

The documentation read "ReadFile reads the named file and returns
the contents. A successful call returns err == nil, not err == EOF.
Because ReadFile reads the whole file, it does not treat an EOF
from Read as an error to be reported."

It does not mention anything about wether there is a file handle
that needs to be closed. ... but there is no file handle returned.
I guess there is no need to close anything.

There are two ways of doing this. The simple way is with
ReadFile/1. The other is with Open/1 and Read/1, making sure to
defer the Close/1 on the file handle. I find this idea of having
multiple ways of doing things to be more hard than good. If you
want something out of a file, you are going to need to get a hold
of it through the os using a file handle, which you need to close.
Familiarity and cohesion have more value than convenience.

Although in this case, since there is a loop, I think defer is not
an available option. Software design and convetions can get really
messy.

Also note that the documentation for Open/1 does not mention
anything about closing the file handle:

"Open opens the named file for reading. If successful, methods on
the returned file can be used for reading; the associated file
descriptor has mode O_RDONLY. If there is an error, it will be of
type *PathError."

I find that disheartening, for some reason. It means that there
are things that cannot be learned through the official
documentation.