Setup for validation complete
I have all the contents for the validation extracted. The next
step is to perform the validations.

I am having second thoughts on performing this. After the issues
that I had with value-based validations and since the property
based validation is simple in this case, I see no reason to write
this code; other than as an exercise. That is the only reason why
I am doing it, as an excercise. But at the same time, I am not
confident that my property based approach is correct either. The
property based approach could have edge cases that are not handled.
It is not so simple. This is the reason why I believe it is
crucial to have both types. A small, simple value based test and a
property-based approach. The value based is tailored towards a few
base cases and also has edge cases. If all the value based pass
but the property based ones do not, then there is reason to believe
something is wrong with the property tests.

Another nice property of value based approach is that non-technical
people can contribute. The client could come up with the (input,
output) pairs itself. It is a non-verbal, simple, and decisive
specification straight from the client - I think that is something
that has a lot of value. It is actionable right away.

All that go test requires is that the file end on _test.go. I think
I could do main_value_based_test.go and main_property_based_test.go
and go test would not mind. That way I can keep both test styles
separate.