Files could not be opened
I am not sure why. I should have logged the error as well. I should
have also printed the entire file name to see if there was an error
in that

Also, print debugging is really slow. This is not the right way to
go about developing software. I need to figure out a way to do it
better.