There is output
I got output from the files. I forgot that the order was
alphabetical, so I freaked out when the first case was not the
first displayed. This made me think that the loop had skipped one
file or something. Thus, I need to verify that the number of test
cases being read are the correct ones. I also need to make sure
the outputs are correct.

This is definetely too slow. This is not the right way to develop
software.