All ok
I just remembered that I had already checked that all test case
directories where there. In this contraption, had one file not
been able to open, I would have noticed, since the error handling
logic is in there.

Why would an error like that happen? A typo. A typo in the file
names (like verfication.md) or a typo in the strings in the src
representing the files.

Typos are a big issue when I write software.