Unexpected validations
I got errors in the verification of all files except one. Did I
do somethign wrong with the code or did I do something wrong with
the test cases. If there is something wrong with the test cases,
I think it is best for me to ditch the value based approach and go
into the property based approach. 4/5 wrong. That is a 75% rate of
human error. Is any of this effort even worth it?

The software needs to be specified someway. And I don't think a
whimsical trial-and-error approach of trying to get a "I like the
output" is the way to go either. I think testing is worth it. I
just need to figure out how to do it right.