I made a mistake
I made a mistake in line 22 in case-two. I missed a line. It
should be easy to fix. In this case finding the error was easy.
That is the reason why I chose case-two. For harder cases, I might
have a harder time figuring out where the mistake is. It might
take several iterations between fixing files and running the tests.
I guess I could make a utility to diagnose where the mistakes. That
is, something that could do some diffing. Since missing one line
messes up every line afterwards, the most critical thing would be
to find the first line where the files do not match. It would still
require iterations, but this is the most sensible approach. I guess
finding the exact char location would also be helpful, as in the
case with the em-dashes.

Yeah, I guess that is the right approach. It is an investment. All
the effor that I am putting in right now is effort that I don't
have to spend in the future.