Do not focus on efficiency
I was trying to figure out if there was an Index variant that takes
a start index so that I can iterate line by line without having to
do string manipulations on the input strings, only on the local
strings that I was going to use, but I couldn't find such Index fn
in the strings package. Maybe I didn't look enough, but there is a
split function. It would be much easier to just split the string
into lines and make the comparisons and be over with it.

It is tempting to attempt to perform a state machine that traverses
the lines and checks if every line is equal without doing any string
manipulations to the input strings, but it is critical to deliver
working software first and then work on performance concerns. To be
honest, I was not thinking about performance. I am just familiar
with low-level API and manipulations. It is a way of thinking that
I got from C and has never left me. I like it, but it is not
appropriate right now. Specially since I have already taken too
long with all these testing shenanigans. I haven't even started to
work on the actual formater. In a real software shop, my superior
would be pissed (technical or not). Deliver quick. Fix latter. The
gentle lie that all devs tell themselves. Becuase the powers that
be have already auctioned latter without letting us know. Now is
the best time to get things right, but such thinking is not
profitable. And so we speed our ways into ... what exactly?
Certainly not to a situation were software is safer and easier to
reason about. I am not sure where it is headed. All I know is that
I can't afford not to tag along. I wish I could.