Moved into internal package
I moved the test into an internal package. But I am realizing that
the signature is not right. As I was writting the new signature of
FindLineMismatch I felt uncomfortable that it took *testing.T as an
argument. Now I realize that that is a cumbersome restriction. Line
mismatch should return the line where the mismatch happened. That
is, it must return an integer. It must be a proper utility function,
in that it must be agnostic of the role in plays in the software
that is using it.