Test fails
That was a dumb mistake. I forgot to call the actual test. I don't
like this idea that if a function terminates and there is no Fail
that implies a success. I think it would be better if devs were
explicit about success. In the event that the test finishes without
success or failure, then that is a test error.

I also formated the main file.