Explanation comments
I need to learn how to use the built in docstrings that Go has but
right now I want to finish the goals that I have so far. I am not
even sure if this is the right way to do multiple tests of the
same function in Go. I don't think it is. I just want to get this
code out faster. I am already taking too long.