One more test
As I was working on the implementation of FindLineMismatch I
realized that I had not defined what to return in the event that
the two strings were identical. I also forgot to decide what to
do if one of the strings was empty. In that case I opted to use an
error, but to be honest, I don't know what to do in that case. I
guess it would be reasonable to retun 1, since that is where the
first mismatch is. What would be more useful? To report an error
or to give an extrapolated answer?

In the strings package the only functions taht return an error are
the ones that manipualte an abstraction on top of a string, like a
Reader (an iterator?). All others do not return error. Index/2
uses that tactic. I find the definition of Cut/2 interesting. It is
analogous to returning an error.

On a very tecnical sense, the answer of comparing a non-empty string
and an empty string should be 1. An empty string has only one
sentence, which is the empty string. Since the empty string is
deffernt from any string, and the non-empty string must have at
least one sentence with an empty string, then it follows that the
two differ in the first sentence. Yeah, that makes sense. I need to
add that test next.