one string empty test
Implemented test where one string is empty. As discussed two commits
ago, the result should be 1. I forgot to think of the case where
both strings are empty. In that case the two strings are equal and
the result shoudl be -1.

I have some concerns about strings that do not have a newline
character. So I am going to make some tests about that.

Adding each test manually to the entry point test is error prone
and tedious. Isn't there a better way to do this?