Dbg fix
It is an embarrasing mistake. I forgot how the += operator works.
It took me a couple of tries, but I finally got the debugger
working.

I don't like how Go test is running the tests and I don't like how
VS Code reports errors, but I am monkey patching my way through
TD, so chances are that I am not using the library the way it is
intended.

I can use the current setup to figure out my way through the code.
This is a big improvement over what I could achieve with gedit.
But for running tests and using git, I still prefer the terminal.