Faulty test
While debugging test testOneWordMismatchAtLastLine I realized that I made a mistake. There is an
empty line in the test case.

left:
1 - These strings\n
2 - have an empty line\n
3 - \n
4 - and a mistmatch\n    <-- there is also a typo here
5 - in the last line.

right:
1 - These strings\n
2 - have an empty line\n
3 - \n
4 - and a mistmatch\n    <-- there is also a typo here
5 - in the end.

Therefore the correct answer should be 5, not 4.

This got me thinking: how is strings.Split going to handle the substring "\n\n"?

This is what the dbgr outputs:

[]string len: 5, cap: 5, ["These strings","have an empty line","","and a mistmatch","in the last line."]
                                                               ^
Here it is ----------------------------------------------------|

An empty string. Thus, strings.Split is sufficient for the kind of operations that I want to
perform. This reminds me of when I was reading the documentation for strings.Split and how it
suggested to use strings.Cut is you wanted to have more control over how to segment a string. In all
the string APIs that I have worked before, I don't recall there being a Cut method. I really like
the name: cutting a string of characters. Another surprising think about the strings library is that
there is no Substring method or a Concat method (I just realized that it is probably because there
is the [start:end-excluded] notatnion and + operator for strings).

What is really odd is that the Index function has no starting point. That one is really odd. All
APIs I have used before have one like that, either overloaded or with a default param value.

Looking at https://stackoverflow.com/questions/25837030/find-index-of-a-substring-in-a-string-with-start-index-specified
I see how there is a common sentiment of frustration, although it is straightforward to implement
using a subslice of the string -> someString[startPoint:].Index(charToFind).

The point is, that these naming conventions, expected usage, and API design are all arbitrary. You
have to look for the right tool for the job that you are trying to complete. If the name throws you
off or if there is some edge case where the function did not conform to your expectations (from
reading the docs or from experience), then you just have to look for the right tool or implement one
yourself. That is all there is to it. And testing helps with this.

If I was doing generative, property-based testing, making sense out of what is going on would have
been much more dificult. But I ca do that from value-based testing. Thus, that is something in
favor of value-based testing. It is great for troubleshooting. ... Although I must not ignore the
fact that this test case is wrong. It could be the case that the code is doing the right thing, but
the test is failing. That would really suck. Then again, I am able to reason about all these things
because the test is simple. With a generated test for property-based testing, this would have been
a lot more difficult.

Stepping through the test case, it turns out that the function is correct. What is incorrect is the
test. That sucks. But again, it is an investment. The previous test did help me out in fixing a
value. In that redard, the tests are neutral: they help me once and slowed me down once. Althought
the wrong test indicates how I was not fully contemplating the problem. So it isn't the tests fault.

Now I need to correct the typo, correct a comment, and fix all tests.

I almost forgot to change the error message. I almost forgot to change the 4 to a 5. In the future,
I should assign a symbol for the expected value to avoid this issue. I could change the tests to
reflect this, but I am already behind scheduled. From here on out, if I modify a test I will make
this change, but I will leave all other tests unchanged.