Hide access safety logic
To be honest, I am not sure if this is the best way to handle it.
I think that the case where one string is a substring of the other
is the case that locks this kind of logic, but I am sure there is
a better way to handle accessing out of bounds. I miss Result/Maybe.