"It is not technically needed"
Should I make this distinction? I think based on the description
of the function, it is easy to understand the implementation. It
is a quick check that could speed up the code as well.

There is also something nice about reducing the possible states
that the main problem has to handle. It gives the developer a
stronger, simpler foundation; freeing them from worrying about edge
cases.

I prefer it the way it was before.