Restoring file
And adding a useful comment. I don't know how to explain it
in code. I guess I could run the actual logic, just to emphasize.
Like this:
  if strings.Contains(left, right) || strings.Contains(right, left) {
    return lineNumber
  else {
    panic("this shouldn't have happened!")
  }

I think there is value in explicitly checking that the condition
we expect the program to be in is actually happening and catch any
edge case might, in a rare case, trigger the unreachable code case.
But it might be overkill, right? Can I prove that the only
possible state of the program in that block of code is that one
of the strings is a substring of the other?

Both strings contain at least one line. This means that there must
be at least run of the if block. If the else block has been
reached, that means that one of the strings has been traversed.
The only way for the else block to be reached is if one of the
strings is traversed without ever having mismatching lines. That
means that up to the last line of the traversed string, the two
strings are equal. Meaning, the traversed string is a substring of
the string that has not been traversed.

Quod erat demostratum

Not really. As I was looking at the code, it made me question if
it is true that the if block must run at least once. The hamster
in my head got to work and I started pondering about what values
could trip the algorithm? What about two strings containing only
one word and both words are different? As I understand that won't
trip strings.Split ... but maybe it does. Then I thought of a
really evil example. What about "\n"? A string with a single
newline char. What is the output of strings.Split? Is it a slice
with a single empty string? or is a slice with two empty strings?
One way to interpret strings.Split is to say. Give me whatever is
between the start of the string and the first instance of the
delimeter, whatever is between delimeters, and whatever is beteen
the delimeter and the end of the string, in that order; with the
special case of returning the original string if the delim is not
found. Using that definition, for Split("\n","\n") we could
interpret it as 'what is between the beggining of the string and
the delim?' the empty string. And 'what is in between the delim and
the end of the string?' The empty string. Then,
  Split("\n", "\n") == ["",""]

Makes sense.

But my code handles the case of two strings being equal before it
attempts to do anything more complicated. Then we can let
  left := "\n"
  right := "\n\n"

That would be bad. I did not consider those cases.

But then again, what are the odds that I would encounter such
cases? Would my software be an utter failure if it cannot handle
such cases?

Maybe I can prevent this case if I use SymbolsOnly. If one of the
strings is composed solely of whitespaces and newlines ... ignore
it? Throw an error? Something. Not sure. I need a break.