Added test. Failed test
It is the edge case. I was expecting the test to fail.

Dangggg I got it right!!!!

  Split("\n", "\n") == ["", ""]

What is weird is Split("\n\n", "\n"). These are the outputs:

leftLines: []string len: 2, cap: 2, ["",""]
rightLines: []string len: 0, cap: 824634884484, []

Based on my logic, the result should be

  Split("\n\n", "\n") == ["", "", ""]

What three strings do you need to insert in "%s\n%s\n%s" to turn it into "\n\n"? "","","".

But it gives that weird answer.

... nevermind. I am not sure what happened to my debugger, but somehow I caught the declaration of
rightLines without giving it any value. Just before entering the for loop the debugger reports

rightLines: []string len: 3, cap: 3, ["","",""]

Which means that my intuition is correct.

... and now I can see why the code is failing.

Split("\n", "\n") == ["", ""] does not fit my definition of a line. In this case is should only be
one line, not two. And Split("\n\n", "\n") is two lines and not three. The issue is that to me if
newline is right before end-of-string then there is no line in between those two characters. When
I mention lines, I mean the lines that you would see in a text editor. The idea of this package is
to be a diagnostic tool for the testing that I am doing for w100. YES, I went in a really long
tangent. One way to solve it is to delete the newline char if it is the last char in the string.
That way, Split will match the definition of a line that I need. That is if the other string also
has a newline in the end. Otherwise that marks a line difference.

Case A: both files have a newline at the end

  file 1            |  file 2
first line\n        |first line
second line\n       |second line\n
final line\n        |final line\n

Solution: delete the newline from both

But this case is redundant, since the function already handles when two strings are equal.

Case B: one file has a newline at the end

  file 1            |  file 2
first line\n        |first line
second line\n       |second line\n
final line\n        |final line

Solution: delete the last newline

But what if there are many newline chars at the end of the file?

  file 1            |  file 2
first line\n        |first line
second line\n       |second line\n
final line\n        |final line\n
\n                  |\n
\n                  |\n
\n
\n

If the two strings are exactly the same having newlines at the end wont make a difference. If they
are different, the end won't be reached, thus the 'ghost line' won't be an issue.

This case though ... it is not very relevant. But it is a difference that would be hard to spot. I
was using gedit to edit the files. Gedit has this weird bug that upon opening a file you cannot
scroll to the last line. I've had to add extra lines to get the program unstuck. Not to mention that
it is something that anyone could miss. It is a nasty case. Case B, for example, has, under my
definition, the same number of lines. But it will cause the program to fail. That one is even easier
to make. I need to handle these cases. There is no way around it.

The good thing is that I already know where the issue is: strings.Split does not match the
definition of newline that I need. I am going to have to come up with a better LineSplit. That one
is going to be a whole function on its own. I think I can reuse the tests that I have. Before doing
that I need to add these two test cases.

After running the test I realized that all that refactoring that I did makes it harder to debbug the
code. The issue is that since I am not creating variables, I don't get to see the values with the
debbuger. But the compiler can optimize that away. I am not getting any benefit by not allocating
variables. I should change that.