Added two tests
I added the two tests mentioned in the previous commit. I am
starting to question my sanity. The desicions that I am making
with these edge cases ... I don't think they are right.

This package is a utility that will be used by the diagnostics
utility to write the tests for w100. (Should I be doing TDD!?!?!
 - Is this the right way to do TDD? Probably not).

The goal is of this package is to help correct invalid tests. There
are two properties that the tests of w100 need to satisfy:
     // from symbols package
  1) SymbolsOnly(input) == SymbolsOnly(output)    and
  2) output == verification

I got some failing validations. In case two, output != verification.
Turns out the difference is in line 22. I wanted to write a
function that would tell me where the mistake happened so that I
don't have to eye-ball it. That is it. Requirement 2) is strict.
The two files must be equal. The case can be made that differences
newline characters at the end are not important ... but this is a
file formater. Even though formating empty lines should be trivial,
it is a case that the software needs to handle. Therefore, I must
be more strict in the definition of these edge cases. The two files
must be strictly equal and the goal of this utility is to indicate
where they are different to correct them. How the dev chooses to
correct them is up to the dev. Therefore, I need to repair the two
edge cases that I just wrote and the original edge case. I need to
define rigurously how to define line counts. I will work on it
after I am done with dinner but here are some thoughts:

So far I have writen lines like:

line one\n
line two\n
...

That is, the newline char denotes the end of the current line. But
another way to look at it is to interpret a newline char as the
begining of a new line:

~ line one
\nline two
\nline three
...

Where the count of \n + 1 represents the line number

~      |line one
\n(1st)|line two
\n(2nd)|line three
\n(3rd)|line four
...

While I was trying to come up with the algo for w100 I had this
idea of considering anything between a start point(inclusive) and
the next new line char(exclusive) as a line. That interpretation
is still valid and it prevents the issue of a ghost line.

I don't want to think about this right now. I will continue after
my break.